backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.24
django-cors-headers==4.4.0
djangorestframework==3.15.2
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
